Strangler extraction #18: retro-UI cluster -> zxnu_retro_ui.py
Four small builders, each at its block''s historical position: the
SD-card and Help tabs'' "Retro" pygame-log toggles (build/disable/
persist/toggle + connects), the sidebar sync-activity getter + Unite!/
Remote-Explorer tab-title colour-cycle timers, and the one-time content
disclaimer dialog (re-bound at the call site for the build_tab_ops kwarg
and the startup calls). zxnu_main.py 4,712 -> 4,431 lines. Tripwired in
test_pane_imports.py.
